Create basic electron application.
Creates a basic electron application that simply opens a blank window to
test that everything is working correctly. Updates package.json to
include electron as a dependency. Adds a start script that invokes
electron. Adds basic code in the entry point (index.js) to allow a blank
window to open to test basic functionality.