Add a simple HTML interface.
Adds a the basic layout of an interface in a HTML file - not connected
yet. For students, note that we're not using <input type="file"> in this
version because of special functionality in electron that will allow us
to open file dialogs in a different way than we normally would do, and
will ultimately make our code less complicated and more idiomatic for an
electron application.