Load HTML Interface in App
Alters the electron code to actually load the newly created main.html
file, and render it in a window for the user. For the benefit of
students, this adds a new require for "path" but that module is part of
the standard library for node.js, and therefore doesn't require a npm
install or any changes to our package.json file.